Add DrupalUtil
